A new approach: a custom `FSEvents` implementation on macOS
This approach aims to use a maximum of two `FSEvent` streams per environment (the active one and the “next” one). Seems to work great.

With this in place, we don't need to use the `native-watcher-registry` at all on any platform. The only platform that needs to be economical with native file-watcher objects has solved that problem through another means.